# Replace regex without catastrophic backtracking