@mpalmer 2017-09-27
Thanks for the PR.  For cleanliness, can you provide a branch containing only the one fix under discussion?