@jcmuller 2017-09-27
@mpalmer: Cleaned.

Should I open another PR removing deprecation warnings?