@mpalmer 2017-09-28
> Cleaned.

Thanks.

> Should I open another PR removing deprecation warnings?

Yes please.